detection_model/requirements.txt
Flask==1.1.2
keras==2.8.0
matplotlib==3.5.2
numpy==1.21.5
opencv_python==4.5.5.64
pandas==1.4.3
Pillow==9.2.0
scikit_learn==1.1.2
tensorflow==2.8.0
tqdm==4.64.0
Werkzeug==2.0.3
